- Cleaned up scripts and check for proper root permissions to be able to set dir permissions

+++ private_tangle/cleanup.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-docker-compose run cleanup
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root or with sudo"
+  exit
+fi
+
 docker-compose down --remove-orphans
 
 rm -Rf privatedb
